Models/requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch==2.2.2
torchaudio==2.2.2
openai-whisper==20231117
nltk==3.8.1
requests==2.31.0
werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0
